Organize flow2 structure
